Creates a plotly object with visual specifications of ptt.

Outputs a plotly object.

Usage

ptt_plot(
  d,
  grouping,
  title,
  subtitle = "",
  caption,
  legend_orientation = "auto",
  legend_position = "bottom",
  margin = NA,
  font_color = "#696969",
  grid_color = "#E8E8E8",
  rangeslider = T,
  start_time = NULL,
  isolate_primary = F,
  font_size = 14,
  hovertext = list(rounding = 1, unit = "", extra = ""),
  height = 500,
  ...
)

Arguments

d

Tibble for plotting.

grouping

Tibble column from tibble d to use for grouping.

title, subtitle, caption

labeling for plot

legend_orientation, legend_position

Legend positioning.

margin

List for plot margins if calculated margins do not work.

font_color, grid_color, font_size

Plot background element colors and font sizes.

rangeslider

Determines rangeslider inclusion.

isolate_primary

Separates the first factor in grouping by line width. Use ptt_add_secondary_traces for multiple series with multiple secondary traces.

hovertext

A list describing hovertext items "list(rounding = 1, unit = "%", extra = "(ennuste)")".

height

Height of the plot.

Value

plotly object

Examples